fix(audit): move _system_deps.py into _core/ to clear PS-108b flat-file threshold
citation_style.py (this slice's new top-level module) pushed
src/scitex_writer/ from 15 to 16 flat .py files at root, tripping the
audit's PS-108b src-flat-py-files-over-threshold rule (>15) -- this is
what failed the `audit` CI job and the single pytest failure on all
three matrix legs (tests/develop/test_audit.py::test_audit_all_clean
runs the same audit check under pytest).

Same fix applied to the sibling PR #230 (feat/engine-count-words-python):
group the CI/deploy system-dependency helper into _core/ per the
audit's own suggested categories, restoring the flat count to 15.
Updates the scitex_dev.system_deps entry-point path and the module's
own python -m invocation docstring to match.
