+++ pyproject.toml
@@ -96,7 +96,7 @@ scitex-writer = "scitex_writer"
 scitex-writer = "scitex_writer"
 
 [project.entry-points."scitex_dev.system_deps"]
-scitex-writer = "scitex_writer._system_deps:provide"
+scitex-writer = "scitex_writer._core._system_deps:provide"
 
 [project.urls]
 Homepage = "https://github.com/ywatanabe1989/scitex-writer"
