+++ src/scitex_writer/_core/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+# File: src/scitex_writer/_core/__init__.py
+
+"""Core infrastructure glue (system deps, config, …) kept out of the flat root.
+
+Groups topical single-purpose modules per PS-108b (`general/02_package/
+02_project-structure-src.md`): CI/deploy/system-dependency helpers that don't
+warrant their own top-level subpackage but also shouldn't keep piling up as
+flat files at ``scitex_writer/`` root.
+"""
+
+# EOF
