+++ src/scitex_writer/_core/_system_deps.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-# File: src/scitex_writer/_system_deps.py
+# File: src/scitex_writer/_core/_system_deps.py
 
 """scitex-writer's native system (apt) dependency provider.
 
@@ -13,7 +13,7 @@ Registered as a ``scitex_dev.system_deps`` entry-point so
 uniform ``scitex-writer dev system-deps list|install`` CLI verb wraps it (added
 with the _cli split). Standalone interim emit (no scitex-dev needed)::
 
-    apt-get install -y --no-install-recommends $(python -m scitex_writer._system_deps)
+    apt-get install -y --no-install-recommends $(python -m scitex_writer._core._system_deps)
 
 NOTE: bibtex (texlive-bibtex-extra) is the default natbib bibliography path,
 matching texlive.def. ``biber`` is also included so biblatex-based manuscripts
@@ -28,7 +28,10 @@ from __future__ import annotations
 _PACKAGES: list[tuple[str, str]] = [
     ("texlive-latex-base", "core LaTeX"),
     ("texlive-latex-recommended", "recommended LaTeX packages"),
-    ("texlive-latex-extra", "extra LaTeX packages (pdfpages, tcolorbox, csvsimple, makecell, ...)"),
+    (
+        "texlive-latex-extra",
+        "extra LaTeX packages (pdfpages, tcolorbox, csvsimple, makecell, ...)",
+    ),
     ("texlive-fonts-recommended", "recommended fonts"),
     ("texlive-fonts-extra", "extra fonts"),
     ("texlive-science", "siunitx + science math packages"),
@@ -55,9 +58,9 @@ def provide():
     """Return this leaf's system deps for the ``scitex_dev.system_deps`` group.
 
     ``SystemDepSpec`` is imported lazily from scitex_dev so this module stays
-    importable (and ``python -m scitex_writer._system_deps`` works) even where
-    scitex-dev isn't installed -- the aggregator that calls ``provide()`` runs
-    in an environment that provides scitex_dev.
+    importable (and ``python -m scitex_writer._core._system_deps`` works) even
+    where scitex-dev isn't installed -- the aggregator that calls ``provide()``
+    runs in an environment that provides scitex_dev.
     """
     from scitex_dev.system_deps import SystemDepSpec
